+++ src/pulumi/core-dtap/analytics/databricks/analytics_workspace.py
@@ -2,7 +2,7 @@ from os import getenv
 from pulumi import Output, ResourceOptions
 import pulumi_azure_native as azure_native
 import pulumi_azuread as azuread
-from pulumi_databricks import Provider as DatabricksProvider, databricks
+import pulumi_databricks as databricks
 
 from ingenii_azure_data_platform.databricks import create_cluster
 from ingenii_azure_data_platform.iam import (
@@ -132,12 +132,13 @@ for assignment in workspace_config.get("iam", {}).get("role_assignments", []):
 # ----------------------------------------------------------------------------------------------------------------------
 # ANALYTICS DATABRICKS WORKSPACE -> PROVIDER
 # ----------------------------------------------------------------------------------------------------------------------
-databricks_provider = DatabricksProvider(
+databricks_provider = databricks.Provider(
     resource_name=workspace_name,
     azure_client_id=getenv("ARM_CLIENT_ID", azure_client.client_id),
     azure_client_secret=getenv("ARM_CLIENT_SECRET"),
     azure_tenant_id=getenv("ARM_TENANT_ID", azure_client.tenant_id),
     azure_workspace_resource_id=workspace.id,
+    host=workspace.workspace_url,
 )
 
 # ----------------------------------------------------------------------------------------------------------------------
@@ -184,7 +185,7 @@ if workspace_firewall_config.get("enabled"):
         enabled=True, ip_access_list=workspace_firewall_config.get("ip_access_list", [])
     )
 
-    databricks.IPAccessList(
+    databricks.IpAccessList(
         resource_name=f"{workspace_name}-firewall",
         label="allow_in",
         list_type="ALLOW",
@@ -401,7 +402,7 @@ mounting_role_assignments = [
 # If no storage mounts are defined in the YAML files, we'll not attempt to create any.
 for definition in workspace_config.get("storage_mounts", []):
     if definition["type"] == "datalake_passthrough":
-        databricks.DatabricksMount(
+        databricks.Mount(
             resource_name=f'{workspace_name}-{definition["mount_name"]}',
             name=definition["mount_name"],
             cluster_id=clusters["default"].id,
@@ -427,20 +428,19 @@ for definition in workspace_config.get("storage_mounts", []):
             ),
         )
     else:
-        databricks.DatabricksMount(
+        databricks.Mount(
             resource_name=f'{workspace_name}-{definition["mount_name"]}',
             name=definition["mount_name"],
-            cluster_id=clusters["default"].id,
-            abfs=databricks.DatabricksMountAbfsArgs(
+            abfs=databricks.MountAbfsArgs(
                 client_id=storage_mounts_sp.application_id,
                 client_secret_key=storage_mounts_dbw_password.key,
-                tenant_id=azure_client.tenant_id,
                 client_secret_scope=secret_scope.name,
-                storage_account_name=storage_accounts[
-                    definition["type"]]["account"].name,
                 container_name=definition["container_name"],
-                initialize_file_system=False
+                initialize_file_system=False,
+                storage_account_name=storage_accounts[definition["type"]]["account"].name,
+                tenant_id=azure_client.tenant_id,
             ),
+            cluster_id=clusters["default"].id,
             opts=ResourceOptions(
                 delete_before_replace=True,
                 depends_on=mounting_role_assignments,