+++ src/python/ingenii_azure_data_platform/ingenii_azure_data_platform/databricks.py
@@ -1,5 +1,5 @@
 from pulumi import ResourceOptions
-from pulumi_databricks import databricks
+import pulumi_databricks as databricks
 
 from ingenii_azure_data_platform.utils import generate_resource_name